+++ Obsidian.Nbt/NbtList.cs
@@ -3,7 +3,7 @@ using System.Text;
 
 namespace Obsidian.Nbt;
 
-public class NbtList : INbtTag, IList<INbtTag>
+public sealed class NbtList : INbtTag, IList<INbtTag>
 {
     private readonly List<INbtTag> baseList = [];
 
@@ -15,8 +15,8 @@ public class NbtList : INbtTag, IList<INbtTag>
 
     public NbtTagType Type => NbtTagType.List;
 
-    public string Name { get; set; }
-    public INbtTag Parent { get; set; }
+    public string? Name { get; set; }
+    public INbtTag? Parent { get; set; }
 
     public NbtList(NbtTagType listType, string name = "")
     {
